Add authentication + user validation to remove station mutation
* Add authentication class
* Add user id check

+++ src/schema.py
@@ -168,7 +168,7 @@ class Mutation:
 
         return WeatherStation.from_model(new_weather_station)
 
-    @strawberry.mutation(description="Remove weather station.")
+    @strawberry.mutation(description="Remove weather station.", permission_classes=[IsAuthenticated])
     async def remove_weather_station(self, info: Info[AppContext, Any], resource_id: int) -> RemoveWeatherStationOutput:
         """Delete weather station from DB"""
         async with info.context.db.session() as session:
@@ -180,10 +180,19 @@ class Mutation:
             result = await session.execute(query)
             weather_station_exists = result.scalar()
             if weather_station_exists:
+                auth_user = info.context.request.auth_user
                 query = select(models.WeatherStation).where(models.WeatherStation.station_id == resource_id)
 
                 result = await session.execute(query)
                 weather_station = result.scalar()
+
+                # Check if deleting own station
+                if weather_station.user_id != auth_user.id:
+                    return RemoveWeatherStationOutput(
+                        resource_id=weather_station.station_id,
+                        message="Cannot remove foreign weather station!",
+                        resource_removed=False,
+                    )
 
                 await session.delete(weather_station)
                 return RemoveWeatherStationOutput(
